+++ src/test/kotlin/com/github/victorrentea/livecoding/Test.java
@@ -1,0 +1,4 @@
+package com.github.victorrentea.livecoding;
+
+public class Test {
+}