+++ src/test/live-coding-playground/src/main/java/playground/AddRACToNested.java
@@ -1,0 +1,11 @@
+package playground;
+
+public class AddRACToNested {
+    static class X {
+        private final int x;
+
+//        X(int x) {
+//            this.x = x;
+//        }
+    }
+}