+++ src/test/live-coding-playground/src/main/java/playground/FinalFieldsWOConstructor.java
@@ -1,0 +1,8 @@
+package playground;
+
+public class FinalFieldsWOConstructor {
+    private final int dummy;
+
+
+
+}