+++ src/test/live-coding-playground/src/main/java/playground/TestSlf4j.java
@@ -1,0 +1,23 @@
+package playground;
+
+
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+public class TestSlf4j {
+
+    private final String name;
+
+    private String name3;
+
+    public TestSlf4j(String name) {
+        this.name = name;
+    }
+
+
+    public static void main(String[] args) {
+
+    }
+
+
+}