+++ src/test/live-coding-playground/src/main/java/playground/VisibleForTestingTest.java
@@ -1,0 +1,16 @@
+package playground;
+
+import com.google.common.annotations.VisibleForTesting;
+
+public class VisibleForTestingTest {
+
+    @VisibleForTesting
+    public void methodForTestsOnly() {
+
+    }
+
+
+    public void abuser() {
+        methodForTestsOnly();
+    }
+}