+++ src/test/live-coding-playground/src/test/java/playground/VisibleFroTestingFromTests.java
@@ -1,0 +1,7 @@
+package playground;
+
+public class VisibleFroTestingFromTests {
+    void correct() {
+     new VisibleForTestingTest().methodForTestsOnly();
+    }
+}